+++ src/Ext/Building/Body.h
@@ -39,6 +39,7 @@ public:
 
 		void DisplayGrinderRefund();
 		void ApplyPoweredKillSpawns();
+		bool HasSuperWeapon(int index, bool withUpgrades) const;
 
 		virtual ~ExtData() = default;